requirements.txt
# ============================================================
# Stock ML System — Pinned Dependencies (Production Discipline)
# ============================================================

# --- Data ---
yfinance==0.2.38
pandas==2.2.2
pyarrow==16.1.0

# --- Feature Engineering ---
ta==0.11.0
nltk==3.9.1
vaderSentiment==3.3.2

# --- ML / Training ---
scikit-learn==1.5.1
xgboost==2.0.3
mlflow==2.14.1

# --- API ---
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4

# --- Dashboard ---
streamlit==1.36.0
plotly==5.22.0

# --- Monitoring ---
scipy==1.14.0

# --- Config ---
pyyaml==6.0.1

# --- Utilities ---
requests==2.32.3
python-dateutil==2.9.0
